refactor: history.c 파일 분리
